pyproject.toml
[tool.poetry]
name = "cipher-genius"
version = "0.1.0"
description = "An LLM-Driven Framework for Automated Cryptographic Scheme Generation"
authors = ["CipherGenius Team"]
readme = "README.md"
license = "MIT"
packages = [{include = "cipher_genius", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
openai = "^1.3.0"
anthropic = "^0.30.0"
zhipuai = "^2.1.5"
langchain = "^0.1.0"
langchain-openai = "^0.0.2"
langchain-anthropic = "^0.1.13"
tiktoken = "^0.5.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
httpx = "^0.25.2"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
celery = "^5.3.4"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.2.2"
numpy = "^1.26.2"
pandas = "^2.1.4"
streamlit = "^1.28.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.0"
ruff = "^0.1.8"
mypy = "^1.7.1"
pre-commit = "^3.6.0"
ipython = "^8.18.1"

[tool.poetry.scripts]
cipher-genius = "cipher_genius.cli.main:app"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=cipher_genius --cov-report=html --cov-report=term"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
